+++ src/XRP/network-clients/json-rpc-network-client.ts
@@ -4,12 +4,14 @@ import {
   JsonRpcRequestOptions,
 } from '../shared/rippled-json-rpc-schema'
 import axios, { AxiosInstance, AxiosResponse } from 'axios'
+import { JsonNetworkClientInterface } from './json-network-client-interface'
 
 /**
  * A network client for interacting with the rippled JSON RPC.
  * @see https://xrpl.org/rippled-api.html
  */
-export default class JsonRpcNetworkClient {
+export default class JsonRpcNetworkClient
+  implements JsonNetworkClientInterface {
   private readonly axiosInstance: AxiosInstance
 
   /**
@@ -65,15 +67,17 @@ export default class JsonRpcNetworkClient {
    *
    * @param account The XRPL account to query for trust lines.
    */
-  public async getAccountLines(account: string): Promise<AccountLinesResponse> {
-    // TODO: consider an option for including the 'peer' param in the request, which limits the returned trust lines to only
-    // those shared between the two accounts. (This would have to be an argument to the method here and in i-c-client too.)
+  public async getAccountLines(
+    account: string,
+    peerAccount?: string,
+  ): Promise<AccountLinesResponse> {
     const accountLinesRequest = {
       method: 'account_lines',
       params: [
         {
           account: account,
           ledger_index: 'validated',
+          peer: peerAccount,
         },
       ],
     }